+++ src/main/java/nu/aron/next/NextMojo.java
@@ -8,10 +8,12 @@ import org.apache.maven.plugins.annotations.Mojo;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import static nu.aron.next.Constants.RUN;
+
 /**
  * A do nothing mojo to be able to call nextversion as a goal
  */
-@Mojo(name = "run", defaultPhase = LifecyclePhase.DEPLOY)
+@Mojo(name = RUN, defaultPhase = LifecyclePhase.DEPLOY)
 public class NextMojo extends AbstractMojo {
     static final Logger log = LoggerFactory.getLogger(NextMojo.class);